Fix bug in materials definition
Materials names defined in DDDRCaloTubes_o1_v01.xml where already
defined in the endcap xml file, therefore the barrel geometry was using
the endcap materials. This is a problem in case one wants to
activate/inactivate optical properties for such materials in the endcap
or barrel separately. Fixed by changing names of DDDRCaloTubes
materials. Also, set particle direction to (0,1,0) in the IDEA_o2_vo1.py
file to shoot into the barrel instead of the endcap.